Adding some more examples
