+++ sdk/messaging/azeventgrid/sample.env
@@ -1,0 +1,4 @@
+EVENTGRID_KEY=<eventgrid-key>
+EVENTGRID_ENDPOINT=<eventgrid-endpoint>
+EVENTGRID_TOPIC=<topicname>
+EVENTGRID_SUBSCRIPTION=<subscriptionname>
